requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
distlib==0.3.9
filelock==3.16.1
idna==3.10
platformdirs==4.3.6
requests==2.32.3
urllib3==2.3.0
virtualenv==20.28.1
python-dotenv==1.0.1

